@claremacrae 2024-09-02
Hi Kris

Thanks very much indeed for working on this.

## Speedup

> reducing the startup significantly

Do you have any measurements that you could record here?

## Testing

So I can know where I can focus my time on testing...

What platforms have you tested it in?

Also, please can you record what specific steps you used to test it.

For example, a good thing to test would be, if when you closed the vault with a Tasks search open in Reading mode, when you reopened the vault, are the correct tasks displayed?