@claremacrae 2024-09-02
> (to be validated with an blank plugin and a sample vault, which was not yet done)

Please could you let me know when you have had a chance to do that? Thanks!