@rxkris 2024-09-02
Platforms tested:
- Linux (Ubuntu 24.04), Obsidian 1.7.1
- Android 14, iodeOS, Obsidian 1.7.1

Example Vault: [TaskTest.zip](https://github.com/user-attachments/files/16841336/TaskTest.zip)

# Speedup (main difference in Total app startup -> Workspace)
- Linux: onLoad vs onLayoutReady: steady 0.2s to 0.3s.
- Android: onLoad 3 to 5 seconds and with onLayoutReady: around 0.4s
- Open Note at startup: No noticable difference which page was used as open note for the workspace loading time (onLoad, onLayoutReady, see testing below) 
- No noticable difference if Tasks is the only plugin or more plugins are installed

# Testing
Android: Force Quit Obsidian, Start Obsidian then View the startup time
- Done with a blank note at startup
- Done with a Task overview page with 5 queries

Both load ways show the correct results, timings the same as shown above or the attached screenshots.

# Screenshots

Small Vault, even bigger difference on the Vault attached

![Screenshot_Both](https://github.com/user-attachments/assets/483714dd-2705-470d-9eca-818ca11c9057)