@claremacrae 2024-09-06
Many thanks again @rxkris!

We (both @ilandikov and I) have tested this on Mac - including with full logging enabled to confirm the changes in behaviour.

I'm merging it now, so that I can test it on my vault on mobile.